+++ src/locales/en/footer.js
@@ -8,6 +8,7 @@ export default {
   footerVatNumber: 'VAT number:',
   footerManagement: 'Management',
   footerSocialMediaTitle: 'Social Media',
+  footerHashtag: '&num;linssimmeläpi',
   footerFacebookAlt: 'Facebook Logo',
   footerGithubAlt: 'GitHub Logo',
   footerInstagramAlt: 'Instagram Logo',